fix: base manager
